# docs/api/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/items": {
            "get": {
                "operationId": "CatalogController_list",
                "summary": "List/search items (code ascending, cursor paginated)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "operationId": "CatalogController_create",
                "summary": "Create a LOCAL item (MOTECH_POS overlay; POSI2000)",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateItemDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/items/{code}": {
            "put": {
                "operationId": "CatalogController_update",
                "summary": "Edit an item (local overlay of price/name; POSI2000)",
                "parameters": [
                    {
                        "name": "code",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateItemDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Envelope { data, meta }"
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "CatalogController_byCode",
                "summary": "Item detail: price + per-warehouse available quantity",
                "parameters": [
                    {
                        "name": "code",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/items/barcode/{bc}": {
            "get": {
                "operationId": "CatalogController_byBarcode",
                "summary": "Resolve an item by barcode (detail + price + stock). Weighted scale barcodes (prefix 02, 12 digits) are decoded: the embedded item code resolves the item and data.scanned.quantity carries the weight in kg.",
                "parameters": [
                    {
                        "name": "bc",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/items/{code}/prices": {
            "get": {
                "operationId": "CatalogController_prices",
                "summary": "All price levels for an item (IAS_ITEM_PRICE): every LEV_NO × unit combination — retail/wholesale/promo lists (POS_ITM_PRICE screen).",
                "parameters": [
                    {
                        "name": "code",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/items/{code}/prices/{levNo}": {
            "get": {
                "operationId": "CatalogController_priceAtLevel",
                "summary": "Price for a specific price level (sale-time level selection). Optional ?unit= picks the unit; defaults to the base (smallest) unit.",
                "parameters": [
                    {
                        "name": "code",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "levNo",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "unit",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/items/{code}/units": {
            "get": {
                "operationId": "CatalogController_units",
                "summary": "All units of measure for an item (IAS_ITM_DTL) with conversion factors (P_SIZE), per-unit barcode and price — e.g. حبة/كرتون.",
                "parameters": [
                    {
                        "name": "code",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/categories": {
            "get": {
                "operationId": "CategoriesController_tree",
                "summary": "Category tree: main item groups (GROUP_DETAILS) with sub-groups (IAS_MAINSUB_GRP_DTL) and item counts. Use ?category= on GET /items.",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/categories/item-types": {
            "get": {
                "operationId": "CategoriesController_itemTypes",
                "summary": "Item nature types (ITEM_TYPES: stocked/service)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Motech POS API",
        "description": "NestJS backend (Oracle-first, Clean/Hexagonal)",
        "version": "0.2.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateItemDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "Item business code (I_CODE) — must be unique"
                    },
                    "name": {
                        "type": "string",
                        "description": "Local name (I_NAME)"
                    },
                    "barcode": {
                        "type": "string"
                    },
                    "unit": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "description": "Local price (excl VAT), >= 0"
                    },
                    "vatPercent": {
                        "type": "number",
                        "description": "Local VAT percent"
                    },
                    "inactive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "code"
                ]
            },
            "UpdateItemDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "barcode": {
                        "type": "string"
                    },
                    "unit": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "description": "Local price (excl VAT), >= 0"
                    },
                    "vatPercent": {
                        "type": "number"
                    },
                    "inactive": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}